+++ src/main/kotlin/klib/exceptions/KonfigParseException.kt
@@ -1,0 +1,11 @@
+package klib.exceptions
+
+/**
+ * Custom exception
+ *
+ * @param reason
+ *
+ * @since <NEXT_VERSION>
+ * @author Thomas Obernosterer
+ */
+class KonfigParseException(reason: String) : Exception("KonfigParseException: Cannot parse konf; Reason: $reason")