+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `toJson` extension for Any
+- `println` Top-Level Function
 ### Changed
 ### Deprecated
 ### Removed